requirements.txt
google_api_python_client==2.124.0
google_auth_oauthlib==1.2.0
google-generativeai==0.4.1
protobuf==3.20.3
python-dotenv==1.0.1
streamlit==1.32.2
